docker file changes and yaml file renamed for testimng purpose

+++ src/main/java/com/krishn/patrtiencclin/clinicalapi/clinicalapi/controllers/PatientController.java
@@ -1,9 +1,11 @@
 package com.krishn.patrtiencclin.clinicalapi.clinicalapi.controllers;
-
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,12 +14,14 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.krishn.patrtiencclin.clinicalapi.clinicalapi.models.Patient;
 import com.krishn.patrtiencclin.clinicalapi.clinicalapi.repos.PatientRepository;
 
 @RestController
 @RequestMapping("/patients")
+@CrossOrigin(origins = "*")
 public class PatientController {
 
     @Autowired
@@ -25,40 +29,73 @@ public class PatientController {
 
     // Create a new patient
     @PostMapping
-    public Patient createPatient(@RequestBody Patient patient) {
-        return patientRepository.save(patient);
+    public ResponseEntity<Patient> createPatient(@RequestBody Patient patient) {
+        try {
+            Patient savedPatient = patientRepository.save(patient);
+            return new ResponseEntity<>(savedPatient, HttpStatus.CREATED);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to create patient", e);
+        }
     }
 
     // Get all patients
     @GetMapping
-    public List<Patient> getAllPatients() {
-        return patientRepository.findAll();
+    public ResponseEntity<List<Patient>> getAllPatients() {
+        try {
+            List<Patient> patients = patientRepository.findAll();
+            return new ResponseEntity<>(patients, HttpStatus.OK);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to retrieve patients", e);
+        }
     }
 
     // Get a patient by ID
     @GetMapping("/{id}")
-    public Optional<Patient> getPatientById(@PathVariable Long id) {
-        return patientRepository.findById(id);
+    public ResponseEntity<Patient> getPatientById(@PathVariable Long id) {
+        try {
+            Optional<Patient> patient = patientRepository.findById(id);
+            if (patient.isPresent()) {
+                return new ResponseEntity<>(patient.get(), HttpStatus.OK);
+            } else {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Patient not found");
+            }
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to retrieve patient", e);
+        }
     }
 
     // Update a patient
     @PutMapping("/{id}")
-    public Patient updatePatient(@PathVariable Long id, @RequestBody Patient patientDetails) {
-        Optional<Patient> optionalPatient = patientRepository.findById(id);
-        if (optionalPatient.isPresent()) {
-            Patient patient = optionalPatient.get();
-            patient.setFirstName(patientDetails.getFirstName());
-            patient.setLastName(patientDetails.getLastName());
-            patient.setAge(patientDetails.getAge());
-            return patientRepository.save(patient);
-        } else {
-            throw new RuntimeException("Patient not found with id " + id);
+    public ResponseEntity<Patient> updatePatient(@PathVariable Long id, @RequestBody Patient patientDetails) {
+        try {
+            Optional<Patient> optionalPatient = patientRepository.findById(id);
+            if (optionalPatient.isPresent()) {
+                Patient patient = optionalPatient.get();
+                patient.setFirstName(patientDetails.getFirstName());
+                patient.setLastName(patientDetails.getLastName());
+                patient.setAge(patientDetails.getAge());
+                Patient updatedPatient = patientRepository.save(patient);
+                return new ResponseEntity<>(updatedPatient, HttpStatus.OK);
+            } else {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Patient not found");
+            }
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to update patient", e);
         }
     }
 
     // Delete a patient
     @DeleteMapping("/{id}")
-    public void deletePatient(@PathVariable Long id) {
-        patientRepository.deleteById(id);
+    public ResponseEntity<Void> deletePatient(@PathVariable Long id) {
+        try {
+            if (patientRepository.existsById(id)) {
+                patientRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Patient not found");
+            }
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Unable to delete patient", e);
+        }
     }
 }
